% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubci-package.R
\docType{package}
\name{ubci-package}
\alias{ubci}
\alias{ubci-package}
\title{ubci: Get Cryptocurrency Related Data by Upbit}
\description{
Get 'UBCI' & Price from 'Upbit'.
'ubci' is the abbreviation of 'UpBit Crypto Index' and is the name of the index provided by 'upbit'.
The 'upbit' is one of the cryptocurrency exchange in Korea. The 'ubci' package is a wrapper around the ticker
information and 'ubci' API supplied by 'upbit'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/ubci}
  \item Report bugs at \url{https://github.com/mrchypark/ubci/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com}

}
\keyword{internal}
