% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubci_get.R
\name{ubci_get}
\alias{ubci_get}
\title{get coin price}
\usage{
ubci_get(code, from, to)
}
\arguments{
\item{code}{coin name fom ubci_get_options() link "UPBIT.KRW-BTC".}

\item{from}{Optional for various time series.
A character string representing a start date in
YYYY-MM-DD format. Default is 2000-01-01.}

\item{to}{Optional for various time series.
A character string representing an end date in
YYYY-MM-DD format. Default is system date today.}
}
\value{
Return type is tibble has columes contain code, date, open, high, low, close, volume.
}
\description{
get coin price data from source. Source is \url{https://upbit.com/exchange}.
This page shows the multiple cryptocurrency price.
}
\examples{
\donttest{
btc <- ubci_get(code = "UPBIT.KRW-BTC")
btc

ubci_get(code = "UPBIT.KRW-BTC", from="2018-05-15")
ubci_get(code = "UPBIT.KRW-BTC", to="2018-05-15")
}

}
