% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubci_get_options.R
\name{ubci_get_options}
\alias{ubci_get_options}
\title{get ubci options}
\usage{
ubci_get_options(lang = c("en", "ko"))
}
\arguments{
\item{lang}{language options. Default is en for English. ko is posible for Korean.}
}
\description{
get ubci options from source.
}
\examples{
\donttest{
ubci_get_options()
ubci_get_options("ko")
}

}
