% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubci_index.R
\name{ubci_index}
\alias{ubci_index}
\title{get ubci index}
\usage{
ubci_index(index = "UBMI", from, to)
}
\arguments{
\item{index}{default is UBMI.}

\item{from}{Optional for various time series.
A character string representing a start date in
YYYY-MM-DD format.}

\item{to}{Optional for various time series.
A character string representing an end date in
YYYY-MM-DD format.}
}
\description{
get ubci index data from source.
}
