% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubci_index_options.R
\name{ubci_index_options}
\alias{ubci_index_options}
\title{get ubci index options}
\usage{
ubci_index_options(lang = c("en", "ko"))
}
\arguments{
\item{lang}{language options. Default is en for English. ko is posible for Korean.}
}
\description{
get ubci index options from source.
}
\examples{
\dontrun{
ubci_index_options()
ubci_index_options("ko")
}

}
