% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubci-package.R
\docType{package}
\name{ubci-package}
\alias{ubci}
\alias{ubci-package}
\title{ubci: Get Cryptocurrency Related Data}
\description{
Get UBCI(UpBit Crypto Index) & Price from Upbit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/ubci}
  \item Report bugs at \url{https://github.com/mrchypark/ubci/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com}

}
\keyword{internal}
